backend/requirements.txt
fastapi==0.136.0
uvicorn==0.45.0
pydantic==2.13.3
pydantic-settings==2.14.0
motor==3.7.1
beanie==2.0.0
httpx==0.28.1
python-dotenv==1.2.2
google-genai==1.73.1
twilio==9.10.5
python-multipart==0.0.26
passlib[bcrypt]==1.7.4
python-jose==3.5.0
bcrypt==4.0.1
